+++ src/redux/config/configStore.js
@@ -1,12 +1,11 @@
 import { configureStore } from "@reduxjs/toolkit";
-import join from "../modules/joinSlice"
+import join from "../modules/join/joinSlice";
 import modalSlice from "../modules/modal/modalSlice";
 const store = configureStore({
-  reducer: {
-    join,
-    modalSlice
-
-  },
+	reducer: {
+		join,
+		modalSlice,
+	},
 });
 
 export default store;
